Added repositorys for store and customer in document DB

+++ app/Data_MySQL/repositories/store_repository.py
@@ -2,7 +2,7 @@ from Data.db import session
 from Data.models.stores import Store
 
 
-def store_changes():
+def store_changes(_):
     session.commit()
 
 
@@ -11,13 +11,19 @@ def find_store(keyword):
 
 
 def remove_store(store):
-    session.delete(store)
-    session.commit()
+    if not store.employees:
+        session.delete(store)
+        session.commit()
+        return f"Butiken {store} har tagits bort."
+    else:
+        return (f"Du får inte ta bort butiken {store} eftersom det finns tillhörande anställda. "
+                f"Ta bort eller flytta dessa innan du försöker igen.")
 
 
 def add_store(store):
-    name, street_address, zip_code, city, phone, email = store_data
+    name, street_address, zip_code, city, phone, email = store
     store = Store(name=name, street_address=street_address, zip_code=zip_code, city=city, phone=phone, email=email)
     session.add(store)
     session.commit()
+    return store
